feat(graphql_formatter): format interfaces with comments
