+++ crates/biome_graphql_formatter/src/comments.rs
@@ -1,4 +1,5 @@
 use crate::prelude::*;
+use biome_diagnostics::category;
 use biome_formatter::comments::{
     is_doc_comment, CommentKind, CommentStyle, Comments, SourceComment,
 };
@@ -6,6 +7,7 @@ use biome_formatter::formatter::Formatter;
 use biome_formatter::{write, FormatResult, FormatRule};
 use biome_graphql_syntax::{GraphqlLanguage, TextLen};
 use biome_rowan::SyntaxTriviaPieceComments;
+use biome_suppression::parse_suppression_comment;
 
 pub type GraphqlComments = Comments<GraphqlLanguage>;
 
@@ -62,12 +64,11 @@ pub struct GraphqlCommentStyle;
 impl CommentStyle for GraphqlCommentStyle {
     type Language = GraphqlLanguage;
 
-    fn is_suppression(_text: &str) -> bool {
-        // parse_suppression_comment(text)
-        //     .filter_map(Result::ok)
-        //     .flat_map(|suppression| suppression.categories)
-        //     .any(|(key, _)| key == category!("format"))
-        false
+    fn is_suppression(text: &str) -> bool {
+        parse_suppression_comment(text)
+            .filter_map(Result::ok)
+            .flat_map(|suppression| suppression.categories)
+            .any(|(key, _)| key == category!("format"))
     }
 
     fn get_comment_kind(_comment: &SyntaxTriviaPieceComments<Self::Language>) -> CommentKind {
