+++ crates/biome_graphql_formatter/src/graphql/lists/directive_location_list.rs
@@ -1,6 +1,7 @@
 use crate::prelude::*;
-use biome_formatter::write;
+use crate::utils::list::write_interface_like_list;
 use biome_graphql_syntax::GraphqlDirectiveLocationList;
+
 #[derive(Debug, Clone, Default)]
 pub(crate) struct FormatGraphqlDirectiveLocationList;
 impl FormatRule<GraphqlDirectiveLocationList> for FormatGraphqlDirectiveLocationList {
@@ -10,22 +11,6 @@ impl FormatRule<GraphqlDirectiveLocationList> for FormatGraphqlDirectiveLocation
         node: &GraphqlDirectiveLocationList,
         f: &mut GraphqlFormatter,
     ) -> FormatResult<()> {
-        for (index, element) in node.elements().enumerate() {
-            let node = element.node();
-
-            if index != 0 {
-                write!(f, [space()])?;
-            }
-
-            write!(f, [node.format()])?;
-
-            let trailing_separator = element.trailing_separator()?;
-
-            if let Some(token) = trailing_separator {
-                write![f, [space(), token.format()]]?;
-            }
-        }
-
-        Ok(())
+        write_interface_like_list(node, f)
     }
 }
