+++ crates/biome_graphql_formatter/src/graphql/lists/implements_interface_list.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use biome_formatter::write;
+use crate::utils::list::write_interface_like_list;
 use biome_graphql_syntax::GraphqlImplementsInterfaceList;
 
 #[derive(Debug, Clone, Default)]
@@ -11,22 +11,6 @@ impl FormatRule<GraphqlImplementsInterfaceList> for FormatGraphqlImplementsInter
         node: &GraphqlImplementsInterfaceList,
         f: &mut GraphqlFormatter,
     ) -> FormatResult<()> {
-        for (index, element) in node.elements().enumerate() {
-            let node = element.node();
-
-            if index != 0 {
-                write!(f, [space()])?;
-            }
-
-            write!(f, [node.format()])?;
-
-            let trailing_separator = element.trailing_separator()?;
-
-            if let Some(token) = trailing_separator {
-                write![f, [space(), token.format()]]?;
-            }
-        }
-
-        Ok(())
+        write_interface_like_list(node, f)
     }
 }
