+++ crates/biome_graphql_formatter/tests/specs/prettier/graphql/comments/interfaces.graphql
@@ -1,7 +1,7 @@
-#type Type1
-#implements
-#A & B &
+type Type1 implements A & B &
 # comment 1
-#                 C & D &
+C & D &
 # comment 2
-# E {a: a}
+E {
+	a: a
+}
