+++ crates/biome_graphql_formatter/tests/specs/prettier/graphql/interface/separator-detection.graphql
@@ -1,30 +1,34 @@
 # This file used to test legacy syntax of interfaces that already removed in `graphql` v16
 # Some comments are outdated
-#type Type1 implements A& B& C& D
-# {&&&&&&&&
-# Should separate by `,` not `&`
-#{a: a}
+type Type1 implements A & B & C & D {
+	# {&&&&&&&&
+	# Should separate by `,` not `&`
+	a: a
+}
 
-#type Type2 implements A& B& C& D
-# &&&&{}&&&&
-# Should separate by `,` not `&`
-#{a: a}
+type Type2 implements A & B & C & D {
+	# &&&&{}&&&&
+	# Should separate by `,` not `&`
+	a: a
+}
 
-#type Type3 implements A&
+type Type3 implements A &
 # &&&&&&&& comment line 1
-   # &&&&&&&& comment line 2
-#B& C& D
-#{a: a}
+# &&&&&&&& comment line 2
+B & C & D {
+	a: a
+}
 
-#type Type4 implements A
+type Type4 implements A &
 # &&&&&&&& comment line 1
-#&
-   # &&&&&&&& comment line 2
-#B& C& D
-#{a: a}
+# &&&&&&&& comment line 2
+B & C & D {
+	a: a
+}
 
-#type Type5 implements A
+type Type5 implements A &
 # &&&&&&&& comment line 1
-   # &&&&&&&& comment line 2
-#&B& C& D
-#{a: a}
+# &&&&&&&& comment line 2
+B & C & D {
+	a: a
+}
